# go_backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Smart Waste Management API
  description: |
    RESTful API for an AI & AR-Driven Smart Waste Management Platform.
    
    ## Features
    - **IoT Integration**: Real-time data from ESP32-based smart bins via MQTT
    - **User Management**: Profile and reward point tracking
    - **Driver Management**: Route optimization and task verification
    - **Dashboard**: Company management, pricing rules, and analytics
    - **Valuation Engine**: AI-based waste pricing
    
    ## Authentication
    Authentication endpoints are placeholders. Implement JWT or OAuth2 for production.
  version: 1.0.0
  contact:
    name: Smart Waste Team
    email: support@smartwaste.io
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
...
paths:
...
  /drivers:
    get:
      tags:
        - Drivers
      summary: List all drivers
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of drivers
    post:
      tags:
        - Drivers
      summary: Create a new driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDriverRequest'
      responses:
        '201':
          description: Driver created
  /drivers/{id}:
    get:
      tags:
        - Drivers
      summary: Get driver by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Driver details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverResponse'
    put:
      tags:
        - Drivers
      summary: Update driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDriverRequest'
      responses:
        '200':
          description: Driver updated
  /drivers/{id}/location:
    put:
      tags:
        - Drivers
      summary: Update driver location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLocationRequest'
      responses:
        '200':
          description: Location updated
  /drivers/{id}/routes:
    get:
      tags:
        - Drivers
      summary: Get optimized routes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: optimize_by
          in: query
          schema:
            type: string
            enum: [distance, fill_level]
            default: distance
      responses:
        '200':
          description: Optimized route
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
  /drivers/{id}/verify:
    post:
      tags:
        - Drivers
      summary: Verify task via QR code
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyTaskRequest'
      responses:
        '200':
          description: Task verified
  /drivers/{id}/stats:
    get:
      tags:
        - Drivers
      summary: Get driver performance statistics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Driver statistics

  # Bins
...
components:
  schemas:
...
    CreateDriverRequest:
      type: object
      required:
        - email
        - password
        - full_name
        - phone
        - license_number
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        full_name:
          type: string
        phone:
          type: string
        license_number:
          type: string
        vehicle_type:
          type: string
        vehicle_plate:
          type: string
    UpdateDriverRequest:
      type: object
      properties:
        full_name:
          type: string
        phone:
          type: string
        vehicle_type:
          type: string
        vehicle_plate:
          type: string
        is_available:
          type: boolean
    DriverResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        full_name:
          type: string
        phone:
          type: string
        license_number:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        is_available:
          type: boolean
        total_collections:
          type: integer
        average_rating:
          type: number
    UpdateLocationRequest:
      type: object
      required:
        - latitude
        - longitude
      properties:
        latitude:
          type: number
        longitude:
          type: number
    VerifyTaskRequest:
      type: object
      required:
        - qr_code
        - collection_id
      properties:
        qr_code:
          type: string
        collection_id:
          type: string
          format: uuid
    RouteResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        driver_id:
          type: string
          format: uuid
        waypoints:
          type: array
          items:
            $ref: '#/components/schemas/Waypoint'
        total_distance_km:
          type: number
        estimated_duration_minutes:
          type: integer
        status:
          type: string
    Waypoint:
      type: object
      properties:
        bin_id:
          type: string
          format: uuid
        device_id:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        fill_level:
          type: integer
        order:
          type: integer
...